test=develop

+++ lite/kernels/host/crop_tensor_compute.cc
@@ -95,7 +95,7 @@ REGISTER_LITE_KERNEL(crop_tensor, kHost, kFloat, kAny, crop_tensor_float, def)
     .Finalize();
 
 using crop_tensor_int32 =
-    paddle::lite::kernels::host::CropTensorCompute<int, PRECISION(kInt32)>;
+    paddle::lite::kernels::host::CropTensorCompute<int, PRECISION(kFloat)>;
 REGISTER_LITE_KERNEL(
     crop_tensor, kHost, kFloat, kAny, crop_tensor_int32, int32_precision)
     .BindInput("X",
